move logic of translate directive to base-translate directive
preparation to add a namespace-translate directive
